services/odm/pyproject.toml
[project]
name = "open-data-mcp-api"
version = "0.0.1"
description = "ODM API 서비스를 위한 REST API"
readme = "README.md"
requires-python = ">=3.10"
host = "0.0.0.0"
port = 8000
log_level = "INFO"
root_path = ""

dependencies = [
    "beanie>=2.0.0",
    "bs4>=0.0.2",
    "elasticsearch==8.10.0",
    "fastapi>=0.116.1",
    "fastapi-pagination>=0.12.0",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "pandas>=2.3.1",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.13.2",
    "selenium>=4.34.2",
    "uvicorn>=0.35.0",
]

[tool.ruff]
exclude = [".venv"]
line-length = 80
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
